% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaDMR.R
\name{aaDMR}
\alias{aaDMR}
\title{DMR detection}
\usage{
aaDMR(object, h = 1000, pcutoff = "fdr", betacutoff = NULL, min.cpgs = 2)
}
\arguments{
\item{object}{A \link{CpGsiteAnnotated} object created from \link{cpgsite.annotate}.}

\item{h}{An agglomerate parameter. Any sig. cpgs within \code{h} distance are collapsed to form a DMR.}

\item{pcutoff}{A threshold to determine DMRs. Default is highly recommended.}

\item{min.cpgs}{The minimum number of consecutive CpGs constituting a DMR.}
}
\description{
The main function of this package. Computes a normalized kernel-weight statistic,
\code{s(x)}, against a null comparison to identify significantly differentially (or variable) methylated regions.

\deqn{ S(x_{i}) = Y_{i} + \sum_{\displaystyle\substack{j = 1\\ j \neq i}}^{n}w_{j}(x_{i})Y_{j}
} where
\eqn{ w_{j} = \dfrac{K\left( \frac{x_{j}-x_{i}}{h}\right)}{\sum_{\displaystyle\substack{j =1\\ j \neq i}}^{n}K\left( \frac{x_{j}-x_{i}}{h}\right)}}
}
\details{
The value of h should be chosen with care. We recommend using a maximum value of 1000bp
as the agglomerate parameter. If  \code{h} is too small DMRs might have refers CpGs and hence DMRs will
span shorter gene loci. Also if \code{h} is too large, DMRs will span larger gene loci with a increased risk of Type I errors.
}
