% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeFDR.R
\name{changeFDR}
\alias{changeFDR}
\title{Change the individual CpG FDR thresholding for a CpGsiteAnnotated object}
\usage{
changeFDR(annot, FDR)
}
\arguments{
\item{annot}{A \linkS4class{CpGsiteAnnotated} object created from \link{cpgsite.annotate}.}

\item{FDR}{The desired individual CpG FDR, which will index the rate at which DMRs are called.}
}
\value{
A re-indexed \linkS4class{CpGsiteAnnotated} object.
}
\description{
Takes a \linkS4class{CpGsiteAnnotated} object and a specified \verb{0 < FDR< 1}, and re-indexes the object in order to call DMRs at the specified rate.
}
\details{
The number of CpG sites called as significant by this function will set the post-smoothing threshold for DMR constituents in \link{aaDMR}
}
