% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpgsite.annotate.R
\name{cpgsite.annotate}
\alias{cpgsite.annotate}
\title{Annotate Illumina CpGs with their chromosome position and test statistic based on \code{limma}.
Annotate a matrix or GenomicRatioSet of 450K or EPIC data with probe weights and chromosomal position.}
\usage{
cpgsite.annotate(datatype = c("array", "sequencing"), object, what=c("Beta", "M"),
arraytype=c("EPIC", "450K"), analysis.type = "differential",
                                              design, contrasts = FALSE,
                                              cont.matrix = NULL, fdr = 0.05, coef, ...)
}
\arguments{
\item{datatype}{Specify whether the data is of type "array" or "sequencing". Currently
only "array" works.}

\item{object}{A matrix-like or GRanges-like object with rows equal to the number of CpG sites interrogated and columns equal the number of samples}

\item{what}{Specify whether matrix-like or GRanges-like object contains \eqn{\beta} values or M-values}

\item{arraytype}{450K or EPIC data. Only works for array \code{datatype}.}

\item{analysis.type}{Options include \code{differential} for \code{aaDMR()} to return DMRs;
\code{variability} to return VMRs. \code{variability} option yet to be implemented.}

\item{design}{Study design matrix. Identical context to differential analysis pipeline in \code{limma}. Must have an intercept if contrasts=FALSE. Applies only when analysis.type \%in\% c("differential").}

\item{fdr}{FDR cutoff (Benjamini and Hocheberg(1995)) for which CpG sites are individually declared significant. Used to index default thresholding in aaDMR(). \strong{Highly recommended as the primary thresholding parameter for calling DMRs}.}
}
\value{
A \linkS4class{CpGsiteAnnotated} class
}
\description{
Annotate Illumina CpGs with their chromosome position and test statistic based on \code{limma}.
Annotate a matrix or GenomicRatioSet of 450K or EPIC data with probe weights and chromosomal position.
}
