% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iddmrAllClasses-class.R
\docType{class}
\name{CpGsiteAnnotated-class}
\alias{CpGsiteAnnotated-class}
\title{S4 objects summarizing individual CpG sites after fitting to model.}
\description{
An S4 class that stores output from \code{\link[=cpgsite.annotate]{cpgsite.annotate()}}
}
\section{Slots}{

\describe{
\item{\code{ranges}}{A GRanges object, containing CpG-level information to be passed to \link{aaDMR}. Mcols of this object include:
\itemize{
\item{stat:} {Per-CpG test statistic; t from limma.}
\item{diff:} {Methylation difference/coefficient. In beta space for \link{cpgsite.annotate} output }
\item{ind.fdr:} {False discovery rate calculated on individual CpG sites.}
\item{is.sig:} {Logical determining whether a CpG site is individually significant or not.}
}}
}}

