% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iddmrAllClasses-class.R
\docType{class}
\name{DMResults-class}
\alias{DMResults-class}
\title{Storage object for DMResults}
\description{
An S4 class that stores DMR information as output from \link{aaDMR}.
}
\section{Slots}{

\describe{
\item{\code{DMResults}}{This class has eight slots, summarizing DMR information to be passed to \code{\link{extractRanges}}:
\itemize{
\item{coord:} {DMR coordinates in UCSC style.}
\item{no.cpgs:} {Number of consecutive CpG sites of DMR.}
\item{min_smoothed_fdr:} {Minimum FDR of the smoothed statistic.}
\item{Stouffer:} {p-value based on Stouffer's summary transform of the \strong{individual} CpG FDRs.}
\item{Fisher:} {p-value based on Fisher combined probability transform of the \strong{individual} CpG FDRs}
\item{maxdiff:} {Maximum differential within the DMR.}
\item{meandiff:} {Mean differential across the DMR.}
}}
}}

