% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractRanges.R
\name{extractRanges}
\alias{extractRanges}
\title{Create a GRanges object from \code{\link{aaDMR}} output.}
\usage{
extractRanges(dmroutput, genome = c("hg19", "hg38", "mm10"))
}
\description{
Takes a DMResults object and produces the corresponding GRanges object.
}
\author{
Daniel Alhassan
}
