% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_dependencies.R
\name{load_dependencies}
\alias{load_dependencies}
\title{Load idDMR package dependencies}
\usage{
load_dependencies()
}
\description{
This function loads (and/or installs) dependencies needed
to use the idDMR package.
}
\author{
Daniel Alhassan
}
